==> output/report-bookdown/01-Introduction.Rmd <==

# Introduction 

We conduct a statistical analysis for multi-use recreational trails in the Bridger Mountains to estimate the volume and location of recreation. Headwaters Economics along with the U.S. Forest Service (USFS), Gallatin Valley Snowmobile Association (GVSA), and Montana Fish Wildlife and Parks (FWP) have deployed trail camera counters along a selection of recreational trails in the Bridger Mountains focused on the summer months of 2021. 

The goals of this analysis is to (1) develop a predictive model of recreational trail use using data from trail counters, weather, trail characteristics, and novel data sources provided by Headwaters Economics and (2) apply the statistical model to demonstrate trade offs in predictive accuracy for different applications. The aim is to inform policy/methods adopted by land management agencies (e.g. U.S. Forest Service) and provide trend insights for local decision makers. Auxiliary data provided through partnerships with Strava and AllTrails was explored to examine the predictive capabilities of these data sources. We conducted analysis of pilot data for 21 trails subsections with trail cameras deployed. We used a generalized additive mixture model (`gamm`) approach which allows for estimation of trends as smooth, non-linear functions of time (and other covariates). We applied a framework of hierarchical generalized additive mixture model (HGAMM) [@pedersen2019hierarchical] and various temporal autoregressive structures on the errors to address the spatial nesting and temporal correlation of these data, respectively. We present results for several models explored and examine the prediction estimates and residuals for both in- and out-of-sample trail subsections. We also provide recommendations on sampling effort (i.e. the spatial and temporal spread of trail camera deployment need to provide reliable predictions of trail use over time) for use both in the Bridgers and more generally.

An overview of the available data and covariates is presented in Section \@ref(Data). An overview of the Generalized Additive Model framework, including models it builds upon and extensions within this framework, is covered in Section \@ref(Models). These models are further explored and explained through an application (with R code and a primer on relevant diagnostic tools) to a single trail in Section \@ref(MidCot). The main analysis of all trails (with monitoring via trail-use counters) within the Bridger Mountains is covered in Section \@ref(AllTrailsAnalysis). Trade-offs in prediction accuracy and various trail types is discussed in Section \@ref(TradeOff). Several short investigations covered during the development of this statistical analysis are included in Appendix Sections \@ref(ATDataUtility) and \@ref(Spatial). 

This report is compiled using the following packages within the R statistical programming language [@R-base]: **bookdown** [@R-bookdown], **rmarkdown** [@R-rmarkdown], and **tinytex** [@R-tinytex]. All associated code for analysis and this report may be found on the GitHub repository associated with this project: https://github.com/MLBartley/HE-TrailUse-R.
