==> output/report-bookdown/08-SpatialNetworkGAMM.Rmd <==
# Spatial Network Generalized Additive Mixture Model {#Spatial}

```{r, eval =  T, include = F}
# library(here)
# library(raster) # DEM for map
# library(tidyverse) # data processing and plotting functions
# library(sf) # tidy spatial data and mapping
# library(mgcv) # fit GAMs and GAMMs
# library(archdata) # source for the Oxford Pots data
# library(tidygraph) # tidy network processing
# # devtools::install_github('thomasp85/ggraph') # need dev version of ggraph 
# library(ggraph) # network plots
# # devtools::install_github('nspope/corMLPE', configure.args = "--with-proj-lib=/usr/local/lib/")
library(corMLPE) # network correlation structure for undirected networks
# # library(ggmap) # for geocoding, but not used by default (see below)
# library(ggspatial) # scale bar for maps
# library(maps) # country outlines for maps


#helps to format this document 
#by default the included code chunks do not appear in the report view
# knitr::opts_chunk$set(echo = F, 
#                       eval = F,
#                       include = F,
#                       message = F, 
#                       # cache=TRUE, 
#                       warning=FALSE, 
#                       error=FALSE, 
#                       root.dir = normalizePath(".."))

source(here("scripts/01.01-OrganizeSpatialData.R"))
source(here("scripts/01.02-LoadFunctions.R"))


# runall = T
runall = F
```

## Overview of Potential Model

During our research into potential models for use in this statistical analysis we explored many options that ended up not being used in this report. One such model is presented here as it may prove more useful in future analyses.

We provided an overview of generalized additive models in Section \@ref(Models) and in our analysis use a hierarchical generalized additive model with various temporal correlation structures to fit our trail use data. One drawback of this model is the difficulty of accounting for potential spatial correlation. The Bridger Mountains are home to a complex spatial network of trails some of which we have subdivided into subsections for this analysis. Several trails can be accessed by multiple trailheads, others are more traditional out-and-back or loops. It is a reasonable to assume that trail use would be more similar along subsections of a single trail and possibly between trails that are located closer together. Currently, we account for the random effect of subsection (`subsectionF`) in our models but we are unable to nest subsection within trail or include information about the overall spatial network.

Spatial network GAMs are another extension of the overall GAM framework that allows for the inclusion of a spatial network in a GAM approach that is used to create a spatial correlation structure in the model. Unfortunately, this is not currently an off-the-shelf way to also incorporate a **temporal** correlation structure. A bespoke spatiotemporal correlation structure could potentially created, but that is currently beyond the scope of this analysis. For the Bridger Mountain trail use statistical analysis it was determined that accounting for temporal correlation while including subsection as a random effect was the best way forward. We include an example of how to fit the spatial network model here in case future work may find it useful.

An overview of spatial network GAMs and two case studies with associated R code for analysis and plots are available at: <https://github.com/nick-gauthier/gam-networks>

## Data used

The spatial data GAM requires the data be organized as a spatial network structure that includes (1) a network of "node" locations that mark the start and end locations of the different trail subsections and (2) the response and predictor variables with columns with location "edges" defined by "to" and "from" columns. For many spatial networks, these edges and nodes are easily defined. For example, towns as nodes and roads between them as edges. However, for a recreational trail network the edges are more clearly defined and the placement of nodes can require more thought. While the overall start and end nodes of an out-and-back trail might be easy to also define, other trails that branch and loop can prove more difficult. Spatial delineation of trails into subsections was decided by Headwaters Economics.

## Model Fit

We fit this spatial GAM model using a correlation structure for symmetric relational data from the **corMLPE** package.

```{r runmodel, eval = TRUE, echo=TRUE}
trail_dat <-  net %E>% as_tibble
trail_dat$wday <- as.integer(trail_dat$wday)

if(runall){
  
trail_model <-  gamm(max.camera ~  
                       s(yday, bs="cc") +
                         s(month, k = 7) +
                         s(wday,
                           bs = "cc", k = 7) +
                         s(daily_aqi_value) +
                         s(temp_max_f) +
                         s(precipitation_in, k = 5) +
                        s(totallength_miles) +
                        total_traveltime +
                         max.count,
                      knots = list(yday = c(0,365)),
           # family =  quasipoisson(), # overdispersed count data
           family =  poisson(), # count data
           correlation = corMLPE(form = ~from + to), # symmetric network correlation structure
           # correlation = corAR1(),
           method = 'REML',
           # select = TRUE,
           data = trail_dat, 
           niterPQL = 30
           )

save(trail_model, 
     file=here("output/models/allT_spatial_gamm-corMLPE.rda"), 
     compress='xz')
}

load(file=here("output/models/allT_spatial_gamm-corMLPE.rda"))

```

## Model Diagnostics

The summary output for this `gamm` object shows an adjusted R-sq of 0.57.

```{r summary-spatial, eval = T}
summary(trail_model$gam)
```

A look at plots (see Figures \@ref(fig:tsDiag) and \@ref(fig:mod-acfpacf)) that provide insight into the presence of temporal dependence remaining after the model is fit shows a high degree of temporal autocorrelation.

```{r tsDiag, eval = T, fig.align='center', out.width= "100%", fig.cap="Time series diagnosis plots plots for spatial network GAM."}
allTrail_mod <- trail_dat %>% 
  tidyr::drop_na(c('daily_aqi_value', "totallength_miles"))


with(allTrail_mod, tsDiagGamm(trail_model, timevar = yday,
                          observed = max.camera))
```

```{r mod-acfpacf, eval = T, fig.align='center', out.width = "100%", fig.cap="Times series ACF/pACF diagnosis plots for spatial network GAM."}

layout(matrix(1:2, ncol = 2))
acf(resid(trail_model$lme, type = "normalized"), 
    lag.max = 36, main = "ACF")
pacf(resid(trail_model$lme, type = "normalized"),
     lag.max = 36, main = "pACF")
layout(1)

```

We have included additional diagnostic plots, however no additional issues are apparent beyond the temporal correlation.

```{r mod-draw, eval = T, echo=F, fig.align='center', out.width= "100%", fig.cap="Partial effect plots for spatial network GAM."}
gratia::draw(trail_model$gam)
```

```{r mod-rootogram, eval = T, echo = F, fig.align='center', out.width= "100%", fig.cap="Rootogram for checking for overdispersion in spatial network GAM."}

rg <- gratia::rootogram(trail_model$gam)
gratia::draw(rg)
# 
# sum(residuals(trail_model$gam, type = "pearson")^2) / df.residual(trail_model$gam)

```

```{r mod-predictions, eval = T, echo = F, fig.align='center', out.width= "100%", fig.cap="Assuming a well-fitted model, we would expect all trail subsections exhibiting similar patterns of dispersion around the 1-1 line (and as we are assuming the data is Poisson, the variance around the mean should equal the mean)."}

#add the predicted values from the model 
allTrail_spatial_pred <- transform(allTrail_mod, 
                      mod = predict(trail_model$gam, 
                                      type = "response"))

ggplot(allTrail_spatial_pred, aes(x=mod, y=max.camera)) +
  facet_wrap(~subsectionF) +
  geom_point(alpha=0.1) +
  geom_abline() +
  labs(x="Predicted count", y="Observed count")
```

```{r modelG-predict-spatial, eval = T, echo=FALSE, fig.align='center', out.width="100%", fig.cap="test2"}
gamm_mod_pred <- predict(trail_model$gam, se.fit=TRUE, type = "response")
allTrail_spatial_pred <- transform(allTrail_mod, 
                 mod = gamm_mod_pred$fit, 
                 mod_se = gamm_mod_pred$se.fit)
all_pred <- ggplot(data=allTrail_spatial_pred, aes(x=yday, y=max.camera, group=subsectionF)) +
  facet_wrap(~subsectionF) +
  geom_ribbon(aes(ymin=(mod-2*mod_se),
                  ymax=(mod+2*mod_se)), alpha=0.25) +
  geom_line(aes(y=(mod), col = trailname)) +
  geom_point(aes(color = trailname)) +
  labs(x=expression("Day of Year"),
       y=expression("Trail use"))

## turn this into two graphs with high/low trail use so scales are different and we can see low trail use subsections better

high_pred <- ggplot(data=filter(allTrail_spatial_pred, subsectionF %in% high_use),
                   aes(x=yday, y=max.camera, 
                       group=subsectionF)) +
  facet_wrap(~subsectionF, ncol = 3) +
  geom_ribbon(aes(ymin=(mod-2*mod_se),
                  ymax=(mod+2*mod_se)), alpha=0.25) +
  geom_line(aes(y=(mod), col = trailname)) +
  geom_point(aes(color = trailname), size = 1) +
  labs(x=expression("Day of Year"),
       y=expression("Trail use")) + 
  theme(legend.position="bottom")

# ggsave(filename = here::here("output/figures/high_pred_modG.pdf"),
#        plot = high_pred_G,
#        height = 8.5, 
#        width = 13,
#        dpi=700)

low_pred <- ggplot(data=filter(allTrail_spatial_pred, subsectionF %notin% c("NA", high_use)),
                   aes(x=yday, y=max.camera, 
                       group=subsectionF)) +
  facet_wrap(~subsectionF, ncol = 3) +
  geom_ribbon(aes(ymin=(mod-2*mod_se),
                  ymax=(mod+2*mod_se)), alpha=0.25) +
  geom_line(aes(y=(mod), col = trailname)) +
  geom_point(aes(color = trailname), size = 1) +
  labs(x=expression("Day of Year"),
       y=expression("Trail use")) + 
  theme(legend.position="bottom")

# ggsave(filename = here::here("output/figures/low_pred_modG.pdf"),
#        plot = low_pred_G,
#        height = 8.5, 
#        width = 13,
#        dpi=700)

high_pred
low_pred
```

## Model Compare

This model includes all trail subsections, even ones with overlapping information about Bridger Ridge, for example. So we are unable to directly compare this model with those used with a combined Bridger Ridge subsection as in models G, GS, and GI. However we determined that the amount of temporal autocorrelation was too high to use this model in our statistical analysis. Should a way to implement a combined spatiotemporal correlation structure is developed, we choose to move forward with a `gamm` approach with temporal autocorrelation structure and trail subsection as a random effect.
